requirements.txt
pyzmq
redis
uvloop; platform_system != "Windows"
fastapi 
uvicorn
httpx
websockets